store/requirements.txt
Django>=3.1.4
Pillow>=7.0.0,<7.1.0
djangorestframework>=3.12.2,<3.13.0
djangorestframework-simplejwt>=4.6.0,<4.7.0